requirements.txt
python==3.10.3
Pillow==9.1.1
Plyer==2.0.0
